+++ src/gamelogic/rapidXML/RapidXMLAdapter.hpp
@@ -33,7 +33,7 @@ namespace GameLogic {
         int get_tile_count();
         vector<vector<int>> get_map();
         vector<tuple<size_t, size_t, char*>> get_objects();
-        string get_entity_texture(int x, int y);
+        string get_entity_texture( size_t x, size_t y);
     private:
         vector<char> _buffer;
         xml_document<>* _doc;
